architecture for returning objects instead of hashes
