+++ lib/cubscout/list.rb
@@ -2,36 +2,9 @@ module Cubscout
   class List
     include Enumerable
 
-    class << self
-      def path(the_path)
-        @path = the_path
-      end
-
-      def read_path
-        @path
-      end
-
-      def all(options = {})
-        raise "No path given" unless @path
-
-        first_page = new(Cubscout.connection.get(@path, page: options[:page] || 1, **options).body)
-
-        if options[:page]
-          first_page.items
-        else
-          last_page = first_page.number_of_pages
-
-          other_pages = (2..last_page).to_a.map do |page|
-            new(Cubscout.connection.get(@path, page: page, **options).body)
-          end
-
-          (first_page.items + other_pages.map(&:items)).flatten
-        end
-      end
-    end
-
-    def initialize(raw_payload)
+    def initialize(raw_payload, collection_name)
       @raw_payload = raw_payload
+      @collection_name = collection_name
     end
 
     def page
@@ -51,7 +24,7 @@ module Cubscout
     end
 
     def items
-      raw_payload.dig("_embedded", self.class.read_path)
+      raw_payload.dig("_embedded", collection_name)
     end
 
     def each(&block)
@@ -60,6 +33,6 @@ module Cubscout
 
     private
 
-    attr_reader :raw_payload
+    attr_reader :raw_payload, :collection_name
   end
 end
