+++ lib/cubscout/object.rb
@@ -1,0 +1,5 @@
+module Cubscout
+  class Object
+    include Scopes
+  end
+end
