+++ lib/cubscout/request.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 module Cubscout
-  class Request < Faraday:: Middleware
+  class Request < Faraday::Middleware
     def call(env)
       env.request_headers['Authorization'] = "Bearer #{Config.oauth_token}"
       env.request_headers['Content-Type'] = 'application/json'
